Added ASR reports
