+++ custom/icds_reports/utils.py
@@ -1,0 +1,127 @@
+import json
+import os
+
+from corehq.apps.locations.models import SQLLocation
+from corehq.apps.reports_core.filters import Choice
+from corehq.apps.userreports.models import ReportConfiguration
+from corehq.apps.userreports.reports.factory import ReportFactory
+from corehq.util.couch import get_document_or_not_found
+
+
+class MPRData(object):
+    resource_file = 'resources/block_mpr.json'
+
+
+class ASRData(object):
+    resource_file = 'resources/block_asr.json'
+
+
+class ICDSData(object):
+
+    def __init__(self, domain, filters, report_id):
+        report_config = ReportFactory.from_spec(
+            get_document_or_not_found(
+                ReportConfiguration,
+                domain,
+                report_id
+            )
+        )
+        report_config.set_filter_values(filters)
+        self.report_config = report_config
+
+    def data(self):
+        return self.report_config.get_data()
+
+
+class ICDSMixin(object):
+    has_sections = False
+
+    def __init__(self, config):
+        self.config = config
+
+    @property
+    def subtitle(self):
+        return []
+
+    @property
+    def headers(self):
+        return []
+
+    @property
+    def rows(self):
+        return [[]]
+
+    @property
+    def sources(self):
+        with open(os.path.join(os.path.dirname(__file__), self.resource_file)) as f:
+            return json.loads(f.read())[self.slug]
+
+    @property
+    def selected_location(self):
+        if self.config['location_id']:
+            return SQLLocation.objects.get(
+                location_id=self.config['location_id']
+            )
+
+    @property
+    def awc(self):
+        if self.config['location_id']:
+            return self.selected_location.get_descendants(include_self=True).filter(
+                location_type__name='awc'
+            )
+
+    @property
+    def awc_number(self):
+        if self.awc:
+            return len(
+                [
+                    loc for loc in self.awc
+                    if 'test' not in loc.metadata and loc.metadata.get('test', '').lower() != 'yes'
+                ]
+            )
+
+    def custom_data(self, selected_location, domain):
+        data = {}
+
+        for config in self.sources['data_source']:
+            filters = {}
+            if selected_location:
+                key = selected_location.location_type.name.lower() + '_id'
+                filters = {
+                    key: [Choice(value=selected_location.location_id, display=selected_location.name)]
+                }
+            if 'date_filter_field' in config:
+                filters.update({config['date_filter_field']: self.config['date_span']})
+            if 'filter' in config:
+                for fil in config['filter']:
+                    filters.update({
+                        fil['column']: {
+                            'operator': fil['operator'],
+                            'operand': fil['value']
+                        }
+                    })
+
+            report_data = ICDSData(domain, filters, config['id']).data()
+            for column in config['columns']:
+                column_agg_func = column['agg_fun']
+                column_name = column['column_name']
+                column_data = 0
+                if column_agg_func == 'sum':
+                    column_data = sum([x[column_name] for x in report_data])
+                elif column_agg_func == 'count':
+                    column_data = len(report_data)
+                elif column_agg_func == 'count_if':
+                    statement = '{0} {1} {2}'.format(
+                        column_name,
+                        column['condition']['operator'],
+                        column['condition']['value']
+                    )
+                    column_data = len([val for val in report_data if eval(statement)])
+                elif column_agg_func == 'avg':
+                    values = [x[column_name] for x in report_data]
+                    column_data = sum(values) / (len(values) or 1)
+                column_display = column_name if 'column_in_report' not in column else column['column_in_report']
+                data.update({
+                    column_display: data.get(column_display, 0) + column_data
+                })
+        return data
